+++ games/game2/human.py
@@ -3,7 +3,7 @@ import os
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(__file__))))
 
 import json
-from games.common.character import Character
+from games.common.Character import Character
 
 
 class Human(Character):