+++ games/game4/HeartsBot.py
@@ -1,0 +1,7 @@
+from games.common.Character import Character
+
+
+class HeartsBot(Character):
+
+    def __init__(self, name):
+        super().__init__(name)