+++ games/game4/HeartsGameData.py
@@ -1,0 +1,5 @@
+
+
+class HeartsGameData:
+    def __init__(self):
+        pass