+++ games/game4/HeartsHumanPlayer.py
@@ -1,0 +1,8 @@
+from games.common.Character import Character
+
+
+class HeartsHumanPlayer(Character):
+    def __init__(self, name):
+        super().__init__(name)
+
+