+++ games/game4/main.py
@@ -1,8 +1,0 @@
-
-
-class Game4:
-    def __init__(self):
-        self.result = False
-    def run(self):
-        print("game4 starts")
-