+++ main.py
@@ -1,7 +1,7 @@
 from games.game1.main import Game1
 from games.game2.main import Game2
 from games.game3.CourtPiece import Game3
-from games.game4.main import Game4
+from games.game4.Hearts import Game4
 from player.player import Player
 import random
 import time